Crate spin_serde

Serde serialization helpers

Re-exports

pub use dependencies::DependencyName;

pub use dependencies::DependencyPackageName;

Modules

base64

Base64 (de)serialization

dependencies

Types for working with component dependencies.

id

ID (de)serialization

Structs

FixedStringVersion

FixedStringVersion represents a version string field with a const value.

FixedVersion

FixedVersion represents a version integer field with a const value.

FixedVersionBackwardCompatible

FixedVersion represents a version integer field with a const value, but accepts lower versions during deserialisation.

Type Aliases

KebabId

A “kebab-case” identifier.

LowerSnakeId

A lower-case “snake_case” identifier.

SnakeId

A “snake_case” identifier.